refactor: address review feedback on parser block complexity reduction
- Extract BaseTextParserBlock to deduplicate shared methods
  (_accumulate_parsed_rows, _parse_list_input, _parse_row, generate,
  _clean) between TagParserBlock and RegexParserBlock
- Combine _extract_delimited and _recover_truncated_object into a single
  _extract_or_recover method that counts unclosed delimiters and appends
  the required number of closers, eliminating redundant scanning
- Add symmetric truncation recovery for arrays (previously only objects
  had recovery)
