+++ src/sdg_hub/core/blocks/parsing/__init__.py
@@ -6,12 +6,14 @@ or JSON extraction.
 """
 
 # Local
+from .base_text_parser_block import BaseTextParserBlock
 from .json_parser_block import JSONParserBlock
 from .regex_parser_block import RegexParserBlock
 from .tag_parser_block import TagParserBlock
 from .text_parser_block import TextParserBlock
 
 __all__ = [
+    "BaseTextParserBlock",
     "JSONParserBlock",
     "RegexParserBlock",
     "TagParserBlock",
